+++ packages/wagmi/src/wagmi-provider/__tests__/config-with-eip6963-and-wallets.test.tsx
@@ -17,10 +17,12 @@ vi.mock('wagmi', async () => {
       return {
         connectAsync: async (...args: any[]) => {
           mockConnectAsync(...args);
+          return {};
         },
       };
     },
     useEnsName: () => ({}),
+    useEnsAvatar: () => ({}),
   };
 });