+++ packages/web3/src/connect-modal/style/index.tsx
@@ -220,9 +220,11 @@ const getThemeStyle = (token: ConnectModalToken): CSSInterpolation => {
               width: token.walletListWidth - token.paddingMD * 2,
               left: token.paddingMD,
               height: token.controlHeightLG,
-              backgroundImage: `linear-gradient(to bottom, ${new TinyColor(token.colorBgBase)
+              backgroundImage: `linear-gradient(to bottom, ${new TinyColor(token.colorBgContainer)
                 .setAlpha(0)
-                .toRgbString()}, ${new TinyColor(token.colorBgBase).setAlpha(1).toRgbString()})`,
+                .toRgbString()}, ${new TinyColor(token.colorBgContainer)
+                .setAlpha(1)
+                .toRgbString()})`,
               pointerEvents: 'none',
             },
           },